Move write to stl functionality to a dedicated writer class
